+++ foxes_opt/problems/layout/__init__.py
@@ -4,6 +4,5 @@ Wind farm layout optimization problems.
 
 from .farm_layout import FarmLayoutOptProblem as FarmLayoutOptProblem
 from .regular_layout import RegularLayoutOptProblem as RegularLayoutOptProblem
-from .reggrids_layout import RegGridsLayoutOptProblem as RegGridsLayoutOptProblem
 
 from . import geom_layouts as geom_layouts