+++ foxes_opt/problems/layout/regular_layout.py
@@ -8,7 +8,7 @@ import foxes.variables as FV
 import foxes.constants as FC
 
 
-def _calc_func(valid, P, ct, st_sel):
+def _calc_func(valid, P, ct, algo, mdata, fdata, st_sel):
     """helper function for Calculator turbine model"""
     return (valid, P * valid, ct * valid)
 
@@ -22,6 +22,9 @@ class RegularLayoutOptProblem(FarmVarsProblem):
     ----------
     min_spacing: float
         The minimal turbine spacing
+    initial_values: dict
+        Initial values for opt variables, key:
+        spacing_x, spacing_y, offset_x, offset_y, angle
 
     :group: opt.problems.layout
 
@@ -29,8 +32,8 @@ class RegularLayoutOptProblem(FarmVarsProblem):
 
     SPACING_X = "spacing_x"
     SPACING_Y = "spacing_y"
-    OFFSET_X = "offset_X"
-    OFFSET_Y = "offset_Y"
+    OFFSET_X = "offset_x"
+    OFFSET_Y = "offset_y"
     ANGLE = "angle"
 
     def __init__(
@@ -38,6 +41,7 @@ class RegularLayoutOptProblem(FarmVarsProblem):
         name,
         algo,
         min_spacing,
+        initial_values=None,
         **kwargs,
     ):
         """
@@ -51,12 +55,16 @@ class RegularLayoutOptProblem(FarmVarsProblem):
             The algorithm
         min_spacing: float
             The minimal turbine spacing
+        initial_values: dict, optional
+            Initial values for opt variables, key:
+            spacing_x, spacing_y, offset_x, offset_y, angle
         kwargs: dict, optional
             Additional parameters for `FarmVarsProblem`
 
         """
         super().__init__(name, algo, **kwargs)
         self.min_spacing = min_spacing
+        self.initial_values = initial_values
 
     def initialize(self, verbosity=1, **kwargs):
         """
@@ -88,7 +96,7 @@ class RegularLayoutOptProblem(FarmVarsProblem):
         pmax = b.p_max()
         pmin = b.p_min()
         self._pmin = pmin
-        self._xy0 = 0.5 * (pmin + pmax)
+        self._xy0 = pmin
         self._halfspan = (pmax - pmin) / 2
         self._halflen = np.linalg.norm(self._halfspan)
         self.max_spacing = 2 * (self._halflen + self.min_spacing)
@@ -108,6 +116,37 @@ class RegularLayoutOptProblem(FarmVarsProblem):
             print(f"  n turbines   = {self._nturb}")
             print(f"  turbine mdls = {self._turbine.models}")
 
+        iniv = self.initial_values
+        self.initial_values = {
+            self.SPACING_X: self.min_spacing,
+            self.SPACING_Y: self.min_spacing,
+            self.OFFSET_X: 0.0,
+            self.OFFSET_Y: 0.0,
+            self.ANGLE: 0.0,
+        }
+        if iniv is not None:
+            mins = {
+                v: m for v, m in zip(self.var_names_float(), self.min_values_float())
+            }
+            maxs = {
+                v: m for v, m in zip(self.var_names_float(), self.max_values_float())
+            }
+            for k, v in iniv.items():
+                assert k in self.initial_values, (
+                    f"Invalid initial value key: '{k}', expected one of {list(self.initial_values.keys())}."
+                )
+                assert v >= mins[k], (
+                    f"Initial value for '{k}' too small: {v} < {mins[k]}."
+                )
+                assert v <= maxs[k], (
+                    f"Initial value for '{k}' too large: {v} > {maxs[k]}."
+                )
+                self.initial_values[k] = v
+        if verbosity > 0:
+            print("  initial values:")
+            for k, v in self.initial_values.items():
+                print(f"    {k:12s} = {v}")
+
         if self.farm.n_turbines < self._nturb:
             for i in range(self._nturb - self.farm.n_turbines):
                 ti = len(self.farm.turbines)
@@ -116,7 +155,6 @@ class RegularLayoutOptProblem(FarmVarsProblem):
                 self.farm.turbines[-1].name = f"T{ti}"
         elif self.farm.n_turbines > self._nturb:
             self.farm.turbines = self.farm.turbines[: self._nturb]
-        self.algo.n_turbines = self._nturb
         self.algo.update_n_turbines()
 
         super().initialize(
@@ -154,7 +192,7 @@ class RegularLayoutOptProblem(FarmVarsProblem):
             Initial float values, shape: (n_vars_float,)
 
         """
-        return [self.min_spacing, self.min_spacing, 0.0, 0.0, 0.0]
+        return list(self.initial_values.values())
 
     def min_values_float(self):
         """
@@ -168,13 +206,16 @@ class RegularLayoutOptProblem(FarmVarsProblem):
             Minimal float values, shape: (n_vars_float,)
 
         """
-        return [
-            self.min_spacing,
-            self.min_spacing,
-            -self._halfspan[0] - self.min_spacing,
-            -self._halfspan[1] - self.min_spacing,
-            0.0,
-        ]
+        return np.array(
+            [
+                self.min_spacing,
+                self.min_spacing,
+                0.0,
+                0.0,
+                0.0,
+            ],
+            dtype=config.dtype_double,
+        )
 
     def max_values_float(self):
         """
@@ -188,13 +229,16 @@ class RegularLayoutOptProblem(FarmVarsProblem):
             Maximal float values, shape: (n_vars_float,)
 
         """
-        return [
-            self.max_spacing,
-            self.max_spacing,
-            self._halfspan[0] + self.min_spacing,
-            self._halfspan[1] + self.min_spacing,
-            90.0,
-        ]
+        return np.array(
+            [
+                self.max_spacing,
+                self.max_spacing,
+                1.0,
+                1.0,
+                90.0,
+            ],
+            dtype=config.dtype_double,
+        )
 
     def opt2farm_vars_individual(self, vars_int, vars_float):
         """
@@ -226,23 +270,22 @@ class RegularLayoutOptProblem(FarmVarsProblem):
         a = np.deg2rad(a)
         nax = np.array([np.cos(a), np.sin(a), 0.0], dtype=config.dtype_double)
         nay = np.cross(np.array([0.0, 0.0, 1.0], dtype=config.dtype_double), nax)
-        x0 = self._xy0 + np.array([ox, oy], dtype=config.dtype_double)
 
-        pts = np.zeros((n_states, nx, ny, 2), dtype=config.dtype_double)
+        pts = np.zeros((nx, ny, 2), dtype=config.dtype_double)
         pts[:] = (
-            x0[None, None, None, :]
-            + np.arange(nx)[None, :, None, None] * dx * nax[None, None, None, :2]
-            + np.arange(ny)[None, None, :, None] * dy * nay[None, None, None, :2]
+            self._xy0[None, None, :]
+            + (ox + np.arange(nx)[:, None, None]) * dx * nax[None, None, :2]
+            + (oy + np.arange(ny)[None, :, None]) * dy * nay[None, None, :2]
         )
 
-        pts = pts.reshape(n_states, nx * ny, 2)
-        valid = self.farm.boundary.points_inside(pts.reshape(n_states * nx * ny, 2))
+        pts = pts.reshape(nx * ny, 2)
+        valid = self.farm.boundary.points_inside(pts)
 
-        farm_vars = {
-            FV.X: pts[:, :, 0],
-            FV.Y: pts[:, :, 1],
-            FC.VALID: valid.reshape(n_states, nx * ny).astype(config.dtype_double),
-        }
+        farm_vars = {}
+        for v, d in zip([FV.X, FV.Y, FC.VALID], [pts[:, 0], pts[:, 1], valid]):
+            a = np.zeros((n_states, nx * ny), dtype=config.dtype_double)
+            a[:] = d[None, :]
+            farm_vars[v] = a
 
         return farm_vars
 
@@ -286,34 +329,30 @@ class RegularLayoutOptProblem(FarmVarsProblem):
         naz[..., 2] = 1
         nay = np.cross(naz, nax)
 
-        pts = np.zeros((n_pop, n_states, nx, ny, 2), dtype=config.dtype_double)
-        pts[:] = self._xy0[None, None, None, None, :]
-        pts[..., 0] += ox[:, None, None, None]
-        pts[..., 1] += oy[:, None, None, None]
-        pts[:] += (
-            np.arange(nx)[None, None, :, None, None]
-            * dx[:, None, None, None, None]
-            * nax[:, None, None, None, :2]
-            + np.arange(ny)[None, None, None, :, None]
-            * dy[:, None, None, None, None]
-            * nay[:, None, None, None, :2]
+        pts = np.zeros((n_pop, nx, ny, 2), dtype=config.dtype_double)
+        pts[:] = (
+            self._xy0[None, None, None, :]
+            + (ox[:, None, None, None] + np.arange(nx)[None, :, None, None])
+            * dx[:, None, None, None]
+            * nax[:, None, None, :2]
+            + (oy[:, None, None, None] + np.arange(ny)[None, None, :, None])
+            * dy[:, None, None, None]
+            * nay[:, None, None, :2]
         )
 
-        qts = np.zeros((n_pop, n_states, n_turbines, 2))
-        qts[:, :N] = pts.reshape(n_pop, n_states, N, 2)
+        qts = np.zeros((n_pop, n_turbines, 2))
+        qts[:, :N] = pts.reshape(n_pop, N, 2)
         del pts
 
         valid = self.farm.boundary.points_inside(
-            qts.reshape(n_pop * n_states * n_turbines, 2)
-        )
+            qts.reshape(n_pop * n_turbines, 2)
+        ).reshape(n_pop, n_turbines)
 
-        farm_vars = {
-            FV.X: qts[:, :, :, 0],
-            FV.Y: qts[:, :, :, 1],
-            FC.VALID: valid.reshape(n_pop, n_states, n_turbines).astype(
-                config.dtype_double
-            ),
-        }
+        farm_vars = {}
+        for v, d in zip([FV.X, FV.Y, FC.VALID], [qts[:, :, 0], qts[:, :, 1], valid]):
+            a = np.zeros((n_pop, n_states, n_turbines), dtype=config.dtype_double)
+            a[:] = d[:, None, :]
+            farm_vars[v] = a
 
         return farm_vars
 
@@ -346,13 +385,13 @@ class RegularLayoutOptProblem(FarmVarsProblem):
         x = farm_vars[FV.X][0, sel]
         y = farm_vars[FV.Y][0, sel]
 
-        self.farm.turbines = [t for i, t in enumerate(self.farm.turbines) if i in sel]
-        for i, t in enumerate(self.farm.turbines):
+        turbines = [t for i, t in enumerate(self.farm.turbines) if i in sel]
+        for i, t in enumerate(turbines):
             t.xy = np.array([x[i], y[i]], dtype=config.dtype_double)
             t.models = [m for m in t.models if m not in [self.name, self._mname]]
             t.index = i
             t.name = f"T{i}"
-        self.algo.update_n_turbines()
+        self.farm.reset_turbines(self.algo, turbines)
 
         return FarmOptProblem.finalize_individual(
             self, vars_int, vars_float, verbosity=1