+++ CHANGELOG.md
@@ -84,3 +84,13 @@
 - Introducing minimal package versions of dependencies
 
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v0.6](https://github.com/FraunhoferIWES/foxes/commits/v0.6)
+
+## v0.7.0
+
+- Compatibility with *foxes* v1.7.0
+- Moving currently not running or deprecated opt problems to new `scratch` folder: `GeomLayoutGridded`, `GeomLayout`, `RegGridsLayoutOptProblem`
+- Bug fixes:
+  - Bugs fixed that caused error with `RegularLayoutOptProblem`
+  - Bug fixed with `GG` and layout problems
+
+**Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v0.7.0](https://github.com/FraunhoferIWES/foxes/commits/v0.7.0)